Fix dd.pivot_table dtype to be deterministic (#1693)

+++ dask/dataframe/methods.py
@@ -102,5 +102,7 @@ def pivot_sum(df, index, columns, values):
 
 
 def pivot_count(df, index, columns, values):
+    # we cannot determine dtype until concatenationg all partitions.
+    # make dtype deterministic, always coerce to np.float64
     return pd.pivot_table(df, index=index, columns=columns,
-                          values=values, aggfunc='count')
+                          values=values, aggfunc='count').astype(np.float64)
